+++ dbconfig.ts
@@ -3,6 +3,8 @@ import knexfile from './knexfile';
 import knex from 'knex';
 
 const env = process.env.NODE_ENV;
-const config = env ? knexfile[env] : 'development';
+const configKey = env || 'development';
+
+const config = knexfile[configKey];
 const db = knex(config);
 export default db;