+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - The selected outcome on a bid not being displayed correctly.
 - Two bets being created at once, due to the service worker, by disabling the service worker for now.
+- Empty outcomes being allowed when creating a bet.
 
 ## [1.2.0] - 07-03-2025